backend/requirements.txt
# FastAPI and Web Framework
fastapi
uvicorn
pydantic
pydantic-settings

# LLM and AI Services
openai

# RAG and Embeddings
faiss-cpu
pypdf
rank-bm25
tiktoken

# Database and Storage
supabase

# Authentication and Security
passlib
python-multipart

# Data Processing
pandas
numpy

# Utilities
python-dotenv
requests

# Development and Testing
pytest
pytest-asyncio
black
flake8
mypy

# Audio Processing (for speech-to-text if needed)
librosa
soundfile
pydub

# PDF Processing
pymupdf
